+++ src/components/Common/UserPanel/UserPanel.module.css
@@ -1,0 +1,28 @@
+.detailCont{
+    box-shadow: 0 10px 25px 0 rgba(50, 50, 93, .07), 0 5px 15px 0 rgba(0, 0, 0, .07);
+    border: none;
+    margin-bottom: 15px
+}
+.headingsColor{
+    color: #242565;
+}
+.inputProf{
+    margin-bottom: 15px;
+    display: block;
+    width: 100%;
+    height: calc(1.5em + 0.75rem + 2px);
+    padding: 0.375rem 0.875rem;
+    font-size: 0.925rem;
+    font-weight: 400;
+    line-height: 1.5;
+    color: #495057;
+    background-color: #fff;
+    background-clip: padding-box;
+    border: 1px solid #DBDFE4;
+    border-radius: 0.25rem;
+    box-shadow: inset 0 1px 1px rgb(0 0 0 / 8%);
+    transition: border-color 0.15s ease-in-out, box-shadow 0.15s ease-in-out;
+}
+.rightCont{
+    background-color: #F3F5F6;
+}
