+++ src/components/DashboardAdmin/DashboardAdmin.module.css
@@ -51,7 +51,7 @@
     border-style: none;
 }
 .tabla{
-    width: 70%;
+    width: 50%;
     display: flex;
     justify-content: center;
 }
