+++ src/components/DashboardAdmin/UserDetails/UserDetails.module.css
@@ -4,7 +4,6 @@
     display: flex;
     align-items: center;
     margin-top: 60px;
-
 }
 .subtileUsers {
     margin-left: 200px;
@@ -61,4 +60,7 @@
 }
 .rowsUsers{
     color: #93a3ab;
+}
+.fakeLinkUsers{
+    cursor: pointer;
 }
