+++ src/components/DashboardAdmin/Users/Users.module.css
@@ -17,7 +17,7 @@
 }
 
 .tablaUsers{
-    width: 100%;
+    width: 75%;
     display: flex;
     justify-content: center;
 }
