+++ src/components/Hooks/useRoleProtected.js
@@ -3,7 +3,7 @@ import { useSelector } from 'react-redux';
 import { useEffect } from "react";
 import { useCookies } from "react-cookie";
 
-const useRoleProtected = (role, redirect = '/') => {
+const useRoleProtected = (role = [], redirect = '/') => {
     const user = useSelector(state => state.user);
     const token = useSelector(state => state.token);
     const navigate = useNavigate();
@@ -19,6 +19,8 @@ const useRoleProtected = (role, redirect = '/') => {
                 role.forEach(r => {
                     if (user.Role.name.toLowerCase() === r.toLowerCase()) auth = true;
                 });
+                //if we passed an empty array, it only checks if user is logged in
+                if (role.length === 0) auth = true;
                 if(!auth) navigate(redirect);
             }
         }
