+++ src/components/NavBar/NavBar.jsx
@@ -103,7 +103,7 @@ const NavBar = () => {
                     Seller
                   </Link>
                 ) : null}
-                <Link to="/profile" className={s.menuItem}>
+                <Link to='/profile' className={s.menuItem}>
                   My Profile
                 </Link>
                 <Link
