+++ src/components/Profile/Profile.module.css
@@ -1,0 +1,19 @@
+.conteiner {
+    font-family: Poppins;
+}
+.ContentsubTitleUsers {
+    height: 100px;
+    width: 100%;
+    display: flex;
+    align-items: center;
+    margin-top: 60px;
+}
+.subtileUsers {
+    margin-left: 200px;
+    color: #000;
+    font-family: "Poppins";
+    font-style: normal;
+    font-weight: 700;
+    font-size: 40px;
+    color: #242565
+}
