+++ src/redux/actions/index.js
@@ -19,6 +19,7 @@ import getUser, { GET_USER } from "./getUser";
 import getUsers, { GET_USERS } from "./getUsers";
 import clearUser, { CLEAR_USER } from "./clearUser";
 import updateUser, { UPDATE_USER } from "./updateUser";
+import updateProfile, { UPDATE_PROFILE } from "./updateProfile";
 import clearUpdateErr, { CLEAR_UPDATE_ERR } from "./clearUpdateErr";
 import getArtists, { GET_ARTISTS } from "./getArtists";
 import getTickets, { GET_TICKETS } from "./getTickets";
@@ -48,6 +49,7 @@ export const actions = {
   GET_ARTISTS,
   GET_TICKETS,
   UPDATE_PASSWORD,
+  UPDATE_PROFILE
 };
 
 const actionsCreator = {
@@ -73,7 +75,8 @@ const actionsCreator = {
   clearUpdateErr,
   getArtists,
   getTickets,
-  updatePassword
+  updatePassword,
+  updateProfile
 };
 
 export default actionsCreator;
